+++ lib/jxl/encode_test.cc
@@ -6,6 +6,7 @@
 #include "jxl/encode.h"
 
 #include "gtest/gtest.h"
+#include "jxl/decode.h"
 #include "jxl/encode_cxx.h"
 #include "lib/extras/codec.h"
 #include "lib/jxl/dec_file.h"
@@ -715,27 +716,66 @@ TEST(EncodeTest, JXL_TRANSCODE_JPEG_TEST(JPEGReconstructionTest)) {
   EXPECT_EQ(0, memcmp(decoded_jpeg_bytes.data(), orig.data(), orig.size()));
 }
 
-#if 0
-// This test is commented out until JxlEncoderAddBox is implemented, and is a
-// prototype of JxlEncoderAddBox usage, not a finished test implementation.
+static void ProcessEncoder(JxlEncoder* enc, std::vector<uint8_t>& compressed,
+                           uint8_t*& next_out, size_t& avail_out) {
+  JxlEncoderStatus process_result = JXL_ENC_NEED_MORE_OUTPUT;
+  while (process_result == JXL_ENC_NEED_MORE_OUTPUT) {
+    process_result = JxlEncoderProcessOutput(enc, &next_out, &avail_out);
+    if (process_result == JXL_ENC_NEED_MORE_OUTPUT) {
+      size_t offset = next_out - compressed.data();
+      compressed.resize(compressed.size() * 2);
+      next_out = compressed.data() + offset;
+      avail_out = compressed.size() - offset;
+    }
+  }
+  size_t offset = next_out - compressed.data();
+  compressed.resize(next_out - compressed.data());
+  next_out = compressed.data() + offset;
+  avail_out = compressed.size() - offset;
+  EXPECT_EQ(JXL_ENC_SUCCESS, process_result);
+}
+
 TEST(EncodeTest, BoxTest) {
+  // Tests adding two metadata boxes with the encoder: an exif box before the
+  // image frame, and an xml box after the image frame. Then verifies the
+  // decoder can decode them, they are in the expected place, and have the
+  // correct content after decoding.
   JxlEncoderPtr enc = JxlEncoderMake(nullptr);
   EXPECT_NE(nullptr, enc.get());
 
-  // TODO(lode): create a test image, initialize encoder and options, prepare
-  // next_out and avail_out, and handle status and output buffer after the
-  // JxlEncoderProcessOutput calls below.
-
   EXPECT_EQ(JXL_ENC_SUCCESS, JxlEncoderUseBoxes(enc.get()));
 
-  // Add an early metadata box
-  constexpr const char* exif_test_string = "exif test data";
+  JxlEncoderOptions* options = JxlEncoderOptionsCreate(enc.get(), NULL);
+  size_t xsize = 50;
+  size_t ysize = 17;
+  JxlPixelFormat pixel_format = {4, JXL_TYPE_UINT16, JXL_BIG_ENDIAN, 0};
+  std::vector<uint8_t> pixels = jxl::test::GetSomeTestImage(xsize, ysize, 4, 0);
+  JxlBasicInfo basic_info;
+  jxl::test::JxlBasicInfoSetFromPixelFormat(&basic_info, &pixel_format);
+  basic_info.xsize = xsize;
+  basic_info.ysize = ysize;
+  basic_info.uses_original_profile = false;
+  EXPECT_EQ(JXL_ENC_SUCCESS, JxlEncoderSetBasicInfo(enc.get(), &basic_info));
+  JxlColorEncoding color_encoding;
+  JxlColorEncodingSetToSRGB(&color_encoding,
+                            /*is_gray=*/false);
+  EXPECT_EQ(JXL_ENC_SUCCESS,
+            JxlEncoderSetColorEncoding(enc.get(), &color_encoding));
+
+  std::vector<uint8_t> compressed = std::vector<uint8_t>(64);
+  uint8_t* next_out = compressed.data();
+  size_t avail_out = compressed.size() - (next_out - compressed.data());
+
+  // Add an early metadata box. Also add a valid 4-byte TIFF offset header
+  // before the fake exif data of these box contents.
+  constexpr const char* exif_test_string = "\0\0\0\0exif test data";
   const uint8_t* exif_data = reinterpret_cast<const uint8_t*>(exif_test_string);
-  const size_t exif_size = strlen(exif_test_string);
+  // Skip the 4 zeroes for strlen
+  const size_t exif_size = 4 + strlen(exif_test_string + 4);
   JxlEncoderAddBox(enc.get(), "Exif", exif_data, exif_size, JXL_FALSE);
 
   // Write to output
-  status = JxlEncoderProcessOutput(enc.get(), &next_out, &avail_out);
+  ProcessEncoder(enc.get(), compressed, next_out, avail_out);
 
   // Add image frame
   EXPECT_EQ(JXL_ENC_SUCCESS,
@@ -745,18 +785,66 @@ TEST(EncodeTest, BoxTest) {
   JxlEncoderCloseFrames(enc.get());
 
   // Write to output
-  status = JxlEncoderProcessOutput(enc.get(), &next_out, &avail_out);
+  ProcessEncoder(enc.get(), compressed, next_out, avail_out);
 
   // Add a late metadata box
-  JxlEncoderAddBox("XML ", xml_data, xml_size);
+  constexpr const char* xml_test_string = "<some random xml data>";
+  const uint8_t* xml_data = reinterpret_cast<const uint8_t*>(xml_test_string);
+  size_t xml_size = strlen(xml_test_string);
+  JxlEncoderAddBox(enc.get(), "XML ", xml_data, xml_size,
+                   /*compress_box=*/false);
 
   // Indicate this is the last box
-  JxlEncoderCloseFrames(enc.get());
+  JxlEncoderCloseBoxes(enc.get());
 
   // Write to output
-  status = JxlEncoderProcessOutput(enc.get(), &next_out, &avail_out);
+  ProcessEncoder(enc.get(), compressed, next_out, avail_out);
+
+  // Decode to verify the boxes, we don't decode to pixels, only the boxes.
+  JxlDecoder* dec = JxlDecoderCreate(nullptr);
+  EXPECT_NE(nullptr, dec);
+
+  EXPECT_EQ(JXL_DEC_SUCCESS,
+            JxlDecoderSubscribeEvents(dec, JXL_DEC_FRAME | JXL_DEC_BOX));
+
+  JxlDecoderSetInput(dec, compressed.data(), compressed.size());
+  JxlDecoderCloseInput(dec);
+
+  std::vector<uint8_t> dec_exif_box(exif_size);
+  std::vector<uint8_t> dec_xml_box(xml_size);
+
+  for (bool post_frame = false;;) {
+    JxlDecoderStatus status = JxlDecoderProcessInput(dec);
+    if (status == JXL_DEC_ERROR) {
+      FAIL();
+    } else if (status == JXL_DEC_SUCCESS) {
+      EXPECT_EQ(0, JxlDecoderReleaseBoxBuffer(dec));
+      break;
+    } else if (status == JXL_DEC_FRAME) {
+      post_frame = true;
+    } else if (status == JXL_DEC_BOX) {
+      // Since we gave the exif/xml box output buffer of the exact known correct
+      // size, 0 bytes should be released. Same when no buffer was set.
+      EXPECT_EQ(0, JxlDecoderReleaseBoxBuffer(dec));
+      JxlBoxType type;
+      EXPECT_EQ(JXL_DEC_SUCCESS, JxlDecoderGetBoxType(dec, type, true));
+      if (!memcmp(type, "Exif", 4)) {
+        // This box should have been encoded before the image frame
+        EXPECT_EQ(false, post_frame);
+        JxlDecoderSetBoxBuffer(dec, dec_exif_box.data(), dec_exif_box.size());
+      } else if (!memcmp(type, "XML ", 4)) {
+        // This box should have been encoded after the image frame
+        EXPECT_EQ(true, post_frame);
+        JxlDecoderSetBoxBuffer(dec, dec_xml_box.data(), dec_xml_box.size());
+      }
+    } else {
+      FAIL();  // unexpected status
+    }
+  }
+
+  EXPECT_EQ(0, memcmp(exif_data, dec_exif_box.data(), exif_size));
+  EXPECT_EQ(0, memcmp(xml_data, dec_xml_box.data(), xml_size));
 }
-#endif
 
 #if JPEGXL_ENABLE_JPEG  // Loading .jpg files requires libjpeg support.
 TEST(EncodeTest, JXL_TRANSCODE_JPEG_TEST(JPEGFrameTest)) {